requirements.txt
openai~=0.27.7
python-dotenv~=0.21.1